Rafactory and Fixed bug
