--- .idea/dictionaries/default_user.xml
@@ -1,10 +1,0 @@
-<component name="ProjectDictionaryState">
-  <dictionary name="default.user">
-    <words>
-      <w>anual</w>
-      <w>página</w>
-      <w>thiago</w>
-      <w>usuário</w>
-    </words>
-  </dictionary>
-</component>
