--- .idea/dictionaries/joao_gustavo.xml
@@ -1,69 +1,0 @@
-<component name="ProjectDictionaryState">
-  <dictionary name="joao.gustavo">
-    <words>
-      <w>adquirir</w>
-      <w>ainda</w>
-      <w>alterações</w>
-      <w>apenas</w>
-      <w>assinar</w>
-      <w>assinatura</w>
-      <w>atividades</w>
-      <w>ativo</w>
-      <w>atual</w>
-      <w>atualizadas</w>
-      <w>atualizar</w>
-      <w>ação</w>
-      <w>cancelado</w>
-      <w>cancelar</w>
-      <w>caractere</w>
-      <w>caracteres</w>
-      <w>certeza</w>
-      <w>cobrado</w>
-      <w>cobrança</w>
-      <w>confirmando</w>
-      <w>confirmar</w>
-      <w>contratar</w>
-      <w>criado</w>
-      <w>deletar</w>
-      <w>deseja</w>
-      <w>desfeita</w>
-      <w>deve</w>
-      <w>documento</w>
-      <w>documentos</w>
-      <w>encontrada</w>
-      <w>erro</w>
-      <w>este</w>
-      <w>expirado</w>
-      <w>fatura</w>
-      <w>faturamento</w>
-      <w>finaliza</w>
-      <w>funcionalidade</w>
-      <w>gerado</w>
-      <w>inativo</w>
-      <w>incompleto</w>
-      <w>informações</w>
-      <w>invoces</w>
-      <w>limite</w>
-      <w>menos</w>
-      <w>nenhuma</w>
-      <w>novamente</w>
-      <w>ocultar</w>
-      <w>pelo</w>
-      <w>pode</w>
-      <w>poderá</w>
-      <w>processar</w>
-      <w>próxima</w>
-      <w>próximo</w>
-      <w>reativar</w>
-      <w>retomar</w>
-      <w>salvar</w>
-      <w>senha</w>
-      <w>sucesso</w>
-      <w>tente</w>
-      <w>tipo</w>
-      <w>usar</w>
-      <w>voltar</w>
-      <w>válido</w>
-    </words>
-  </dictionary>
-</component>
