--- .idea/dictionaries/project.xml
@@ -1,26 +1,0 @@
-<component name="ProjectDictionaryState">
-  <dictionary name="project">
-    <words>
-      <w>alterar</w>
-      <w>altura</w>
-      <w>ativa</w>
-      <w>atualizando</w>
-      <w>carregando</w>
-      <w>carregar</w>
-      <w>cnpj</w>
-      <w>copiado</w>
-      <w>copiar</w>
-      <w>creted</w>
-      <w>código</w>
-      <w>encontradas</w>
-      <w>escuro</w>
-      <w>faturas</w>
-      <w>gerenciar</w>
-      <w>momento</w>
-      <w>nextjs</w>
-      <w>perfil</w>
-      <w>pricingpage</w>
-      <w>renovar</w>
-    </words>
-  </dictionary>
-</component>
