--- .idea/inspectionProfiles/Project_Default.xml
@@ -1,22 +1,0 @@
-<component name="InspectionProjectProfileManager">
-  <profile version="1.0">
-    <option name="myName" value="Project Default" />
-    <option name="scopesOrder">
-      <list>
-        <option value="All Changed Files" />
-        <option value="Open Files" />
-        <option value="Project Files" />
-        <option value="Scratches and Consoles" />
-        <option value="Tests" />
-        <option value="tailwind.config.js" />
-      </list>
-    </option>
-    <inspection_tool class="DuplicatedCode" enabled="true" level="WEAK WARNING" enabled_by_default="true">
-      <Languages>
-        <language minSize="230" name="TypeScript" />
-      </Languages>
-    </inspection_tool>
-    <inspection_tool class="Eslint" enabled="true" level="WARNING" enabled_by_default="true" />
-    <inspection_tool class="JSPotentiallyInvalidTargetOfIndexedPropertyAccess" enabled="false" level="WARNING" enabled_by_default="false" />
-  </profile>
-</component>
