+++ apps/server/package.json
@@ -36,9 +36,7 @@
     "crypto": "^1.0.1",
     "express": "^5.1.0",
     "jsonwebtoken": "^9.0.2",
-    "mailersend": "^2.5.0",
-    "mercadopago": "^2.4.0",
-    "nodemailer": "^6.10.1",
+    "nodemailer": "^7.0.2",
     "openai": "^4.95.0",
     "stripe": "^18.0.0",
     "twilio": "^5.5.2"
